@@ -1,77 +1,82 @@
 {
-    "name": "formie",
+    "name": "formie-cp",
     "version": "1.0.0",
     "private": true,
-    "browserslist": [
-        "defaults",
-        "Edge >= 15",
-        "Chrome >= 29",
-        "Safari >= 9",
-        "ie >= 11",
-        "Firefox >= 28",
-        "iOS >= 9.1",
-        "Android >= 4.4"
-    ],
     "scripts": {
-        "dev": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js || true",
-        "watch": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --watch --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js || true",
-        "prod": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js || true"
+        "dev": "npm run forms & npm run cp & npm run frontend",
+        "build": "npm run forms-build & npm run cp-build & npm run frontend-build",
+        "forms": "cd ./src/web/assets/forms && vite --port 4000",
+        "forms-build": "cd ./src/web/assets/forms && vite build",
+        "cp": "cd ./src/web/assets/cp && vite build --watch",
+        "cp-build": "cd ./src/web/assets/cp && vite build",
+        "frontend": "cd ./src/web/assets/frontend && npx mix watch",
+        "frontend-build": "cd ./src/web/assets/frontend && npx mix"
     },
     "devDependencies": {
+        "@babel/eslint-parser": "^7.17.0",
         "@pixelandtonic/craftui": "~0.3.8",
-        "babel-eslint": "^10.0.3",
-        "copy-webpack-plugin": "^5.1.1",
-        "craftcms-sass": "^3.4.0",
-        "cross-env": "^7.0.2",
-        "enhanced-resolve": "^4.2.0",
-        "eslint": "^7.0.0",
-        "eslint-loader": "^4.0.0",
-        "eslint-plugin-import": "^2.22.0",
-        "eslint-plugin-vue": "^6.2.2",
-        "imagemin-webpack-plugin": "^2.4.2",
-        "laravel-mix": "^5.0.4",
-        "laravel-mix-bundle-analyzer": "^1.0.5",
-        "laravel-mix-eslint-config": "^0.1.7",
-        "laravel-mix-imagemin": "^1.0.3",
-        "laravel-mix-polyfill": "^2.0.0",
-        "laravel-mix-stylelint": "^0.2.0",
-        "lodash": "^4.17.21",
-        "postcss-custom-properties": "^10.0.0",
-        "resolve-url-loader": "^3.1.2",
-        "sass": "^1.26.10",
-        "sass-loader": "^9.0.2",
-        "vue-template-compiler": "^2.6.14",
-        "vuex": "^3.5.1"
+        "@rollup/plugin-node-resolve": "^13.3.0",
+        "@vitejs/plugin-legacy": "^1.8.2",
+        "@vitejs/plugin-vue": "^2.3.2",
+        "autoprefixer": "^10.4.7",
+        "babel-eslint": "^10.1.0",
+        "clean-webpack-plugin": "^4.0.0",
+        "craftcms-sass": "^3.5.7",
+        "eslint": "^8.15.0",
+        "eslint-import-resolver-alias": "^1.1.2",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-vue": "^8.7.1",
+        "laravel-mix": "^6.0.43",
+        "laravel-mix-clean": "^0.1.0",
+        "laravel-mix-polyfill": "^3.0.1",
+        "lodash-es": "^4.17.21",
+        "mini-css-extract-plugin": "^2.6.0",
+        "postcss": "^8.4.13",
+        "postcss-color-function": "^4.1.0",
+        "postcss-custom-properties": "^12.1.7",
+        "rollup-plugin-analyzer": "^4.0.0",
+        "sass": "^1.51.0",
+        "sass-loader": "^12.6.0",
+        "vite": "2.9.8",
+        "vite-plugin-eslint": "^1.6.0",
+        "vite-plugin-imagemin": "^0.6.1",
+        "vuex": "^4.0.2"
     },
     "dependencies": {
-        "@braid/vue-formulate": "^2.4.3",
-        "@yaireo/tagify": "^3.14.3",
-        "axios": "^0.21.2",
-        "core-js": "^3.6.5",
+        "@tiptap/core": "^2.0.0-beta.175",
+        "@tiptap/extension-floating-menu": "^2.0.0-beta.51",
+        "@tiptap/extension-focus": "^2.0.0-beta.40",
+        "@tiptap/extension-highlight": "^2.0.0-beta.33",
+        "@tiptap/extension-image": "^2.0.0-beta.27",
+        "@tiptap/extension-link": "^2.0.0-beta.37",
+        "@tiptap/extension-subscript": "^2.0.0-beta.10",
+        "@tiptap/extension-superscript": "^2.0.0-beta.10",
+        "@tiptap/extension-text-align": "^2.0.0-beta.29",
+        "@tiptap/extension-underline": "^2.0.0-beta.23",
+        "@tiptap/starter-kit": "^2.0.0-beta.184",
+        "@tiptap/vue-3": "^2.0.0-beta.91",
+        "@yaireo/tagify": "^4.12.0",
+        "click-outside-vue3": "^4.0.1",
         "expression-language": "^1.1.4",
-        "flat": "^5.0.0",
-        "flatpickr": "^4.6.9",
-        "formdata-polyfill": "^3.0.20",
-        "intl-tel-input": "^17.0.8",
-        "locutus": "^2.0.15",
-        "md5-hex": "^3.0.1",
+        "flat": "^5.0.2",
+        "flatpickr": "^4.6.13",
+        "floating-vue": "^2.0.0-beta.16",
+        "formdata-polyfill": "^4.0.10",
+        "install": "^0.13.0",
+        "intl-tel-input": "^17.0.16",
+        "locutus": "^2.0.16",
+        "md5-hex": "^4.0.0",
         "mobile-drag-drop": "^2.3.0-rc.2",
+        "npm": "^8.9.0",
         "pell": "^1.0.6",
         "places.js": "^1.19.0",
-        "portal-vue": "^2.1.7",
-        "qs": "^6.9.4",
-        "sanitize-html": "^2.3.2",
-        "signature_pad": "^3.0.0-beta.4",
-        "sortablejs": "^1.10.2",
-        "tippy.js": "^6.3.1",
-        "tiptap": "^1.29.0",
-        "tiptap-extensions": "^1.31.0",
-        "v-tooltip": "^2.0.3",
-        "vue-clickaway": "^2.2.2",
-        "vue-drag-drop": "^1.1.4",
-        "vue-final-modal": "^0.21.7",
-        "vue-multiselect": "^2.1.6",
-        "vue-simple-markdown": "^1.1.4",
-        "vue-unique-id": "^3.0.0"
+        "qs": "^6.10.3",
+        "sanitize-html": "^2.7.0",
+        "signature_pad": "^4.0.4",
+        "sortablejs": "^1.15.0",
+        "tippy.js": "^6.3.7",
+        "vue": "^3.2.33",
+        "vue-final-modal": "^3.4.3",
+        "vue-mitt": "^0.0.1"
     }
 }
