@@ -43,7 +43,7 @@ abstract class AddressProvider extends Integration
     {
         $handle = StringHelper::toKebabCase(static::displayName());
 
-        return Craft::$app->getAssetManager()->getPublishedUrl("@verbb/formie/web/assets/addressproviders/dist/img/{$handle}.svg", true);
+        return Craft::$app->getAssetManager()->getPublishedUrl("@verbb/formie/web/assets/cp/dist/img/addressproviders/{$handle}.svg", true);
     }
 
     /**
