@@ -41,7 +41,7 @@ abstract class Captcha extends Integration
     {
         $handle = StringHelper::toKebabCase($this->getHandle());
 
-        return Craft::$app->getAssetManager()->getPublishedUrl("@verbb/formie/web/assets/captchas/dist/img/{$handle}.svg", true);
+        return Craft::$app->getAssetManager()->getPublishedUrl("@verbb/formie/web/assets/cp/dist/img/captchas/{$handle}.svg", true);
     }
 
     public function getFormSettingsHtml($form): string
