@@ -2,8 +2,7 @@
 namespace verbb\formie\web\assets\forms;
 
 use craft\web\AssetBundle;
-use craft\web\assets\cp\CpAsset;
-use craft\web\assets\vue\VueAsset;
+use craft\web\assets\cp\CpAsset as CraftCpAsset;
 
 use verbb\base\assetbundles\CpAsset as VerbbCpAsset;
 
@@ -14,21 +13,26 @@ class FormsAsset extends AssetBundle
 
     public function init(): void
     {
-        $this->sourcePath = __DIR__ . '/dist/';
+        $this->sourcePath = '@verbb/formie/web/assets/cp/dist';
 
         $this->depends = [
             VerbbCpAsset::class,
-            CpAsset::class,
-            VueAsset::class,
+            CraftCpAsset::class,
         ];
 
-        $this->js = [
-            'js/main.js',
-        ];
+        // $this->depends = [
+        //     VerbbCpAsset::class,
+        //     CpAsset::class,
+        //     VueAsset::class,
+        // ];
 
-        $this->css = [
-            'css/style.css',
-        ];
+        // $this->js = [
+        //     'js/main.js',
+        // ];
+
+        // $this->css = [
+        //     'css/style.css',
+        // ];
 
         parent::init();
     }
