@@ -189,7 +189,7 @@ export default {
             this.selectedKey = this.existingFields[0].key;
         }
 
-        this.$events.$on('formie:page-selected', pageIndex => {
+        this.$events.on('formie:page-selected', pageIndex => {
             this.pageIndex = pageIndex;
         });
     },
