@@ -114,13 +114,13 @@ export default {
     },
 
     created() {
-        this.$events.$on('formie:dragging-active', this.draggingActive);
-        this.$events.$on('formie:dragging-inactive', this.draggingInactive);
+        this.$events.on('formie:dragging-active', this.draggingActive);
+        this.$events.on('formie:dragging-inactive', this.draggingInactive);
     },
 
     beforeDestroy() {
-        this.$events.$off('formie:dragging-active', this.draggingActive);
-        this.$events.$off('formie:dragging-inactive', this.draggingInactive);
+        this.$events.off('formie:dragging-active', this.draggingActive);
+        this.$events.off('formie:dragging-inactive', this.draggingInactive);
     },
 
     methods: {
@@ -166,7 +166,7 @@ export default {
             }
 
             // Reset the state
-            this.$events.$emit('formie:dragging-inactive');
+            this.$events.emit('formie:dragging-inactive');
 
             // Is this a pill? If so, we need to insert
             const isPill = (data.trigger === 'pill');
